Add version information including commit id
