+++ bam-cmd/src/main/java/org/victorchang/QnameCommand.java
@@ -1,27 +1,15 @@
 package org.victorchang;
 
-import htsjdk.samtools.SAMSequenceRecord;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
-import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.Option;
-import picocli.CommandLine.Parameters;
 
 import java.io.IOException;
-import java.nio.file.Files;
+import java.io.InputStream;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.Callable;
-
-import static java.util.stream.Collectors.partitioningBy;
-import static java.util.stream.Collectors.toSet;
-import static org.victorchang.QnameCommand.LOG;
+import java.util.Properties;
 
 @Command(
         subcommands = {
@@ -30,10 +18,13 @@ import static org.victorchang.QnameCommand.LOG;
         },
         name = "atlantool",
         header = "A command line tool for viewing records in a BAM file by QNAME",
-        description = "This tool provides option to create an index for a BAM file based on QNAME and then search records based on QNAME."
+        description = "This tool provides option to create an index for a BAM file based on QNAME and then search records based on QNAME.",
+        versionProvider = CommitVersionProvider.class
 )
 public class QnameCommand {
-    static final Logger LOG = LoggerFactory.getLogger(QnameCommand.class);
+    @CommandLine.Option(names = { "-V", "--version" }, versionHelp = true,
+            description = "Print version information and exit")
+    private boolean versionRequested;
 
     public static void main(String[] args) {
         int exitCode = new CommandLine(new QnameCommand()).execute(args);
@@ -50,244 +41,23 @@ public class QnameCommand {
     }
 }
 
-@Command(name = "index",
-        header = "Create a QNAME based index for a BAM file",
-        description = "Create an index for a BAM file based on QNAME. The index can be used later to fetch records from the BAM" +
-                " file efficiently using the `view` sub command.")
-class IndexCommand implements Callable<Integer> {
-    @Parameters(paramLabel = "bam-file", description = "Path to the BAM file")
-    private Path bamPath;
+class CommitVersionProvider implements CommandLine.IVersionProvider {
 
-    @Option(names = {"-i", "--index-path"},
-            description = "Directory to store index files. By default uses a directory name that starts with the BAM file name (so stored next to it)")
-    private Path indexDirectory;
-
-    @Option(names = "--thread-count",
-            description = "Number of threads used for sorting",
-            defaultValue = "1")
-    private int threadCount;
-
-    @Option(names = "--sort-buffer-size",
-            description = "Maximum number of records per buffer used for sorting",
-            defaultValue = "500000")
-    private int sortBufferSize;
-
-    @Option(names = {"-l", "--limit-bytes"},
-            description = "Only read and index first given bytes")
-    private long bytesLimit;
-
-    @Option(names = {"-t", "--temporary-path"},
-            description = "Directory to store temporary files for sorting. By default uses the index path")
-    private Path tempDirectory;
-
-    @Option(names = {"-v", "--verbose"},
-            description = "Switch on verbose output", defaultValue = "false")
-    private boolean verbose;
-
-    @Option(names = {"--force"},
-            description = "Overwrite existing index",
-            defaultValue = "false")
-    private boolean force;
-
-    @Option(names = {"--compression"},
-            description = "Compression level (1 to 9). 1 = faster but bigger index file size, 9 = slower but smaller index file size",
-            defaultValue = "6")
-    int compressionLevel;
+    private final static String BASE_VERSION = "1.0";
+    private static final Logger LOG = LoggerFactory.getLogger(QnameCommand.class);
 
     @Override
-    public Integer call() {
-        java.util.logging.Logger.getLogger("")
-                .setLevel(verbose ? java.util.logging.Level.ALL : java.util.logging.Level.SEVERE);
-
-        if (!Files.isRegularFile(bamPath)) {
-            System.err.println(bamPath + " not found.");
-            return -1;
-        }
-        if (!createIndexDirectory(bamPath, force)) {
-            return -1;
-        }
-
-        if (tempDirectory == null) {
-            tempDirectory = indexDirectory;
-        } else if (!createDirectory(tempDirectory)) {
-            return -1;
-        }
-
-        bytesLimit = bytesLimit == 0 ? Long.MAX_VALUE : bytesLimit;
-
-        BamFileReader fileReader = new DefaultBamFileReader(new EfficientBamRecordParser());
-        QnameIndexer indexer = new QnameIndexer(fileReader,
-                new KeyPointerWriter(compressionLevel),
-                new KeyPointerReader(),
-                threadCount,
-                sortBufferSize);
+    public String[] getVersion() {
+        Properties p = new Properties();
         try {
-            LOG.info("Creating index {} using {} threads with sort buffer size of {} records",
-                    bytesLimit == Long.MAX_VALUE ? "" : "for the first " + bytesLimit + " bytes",
-                    threadCount,
-                    sortBufferSize);
-
-            long start = System.nanoTime();
-            indexer.index(indexDirectory, bamPath, tempDirectory, bytesLimit);
-            long finish = System.nanoTime();
-
-            LOG.info("Create index completed in {}", (finish - start) / 1000_000 + "ms");
+            final InputStream resourceAsStream = getClass().getClassLoader().getResourceAsStream("git.properties");
+            p.load(resourceAsStream);
+            final String commitId = p.getProperty("git.commit.id.abbrev");
+            final String buildTime = p.getProperty("git.build.time");
+            return new String[] { "Version: " + BASE_VERSION, "Release: " + commitId, "Release Date: " + buildTime };
         } catch (IOException e) {
-            LOG.error("Failed to index", e);
-            return -1;
-        } finally {
-            indexer.shutDown();
+            LOG.warn("Failed to read version: ", e);
         }
-        return 0;
-    }
-
-    private boolean createIndexDirectory(Path bamPath, boolean force) {
-        if (indexDirectory == null) {
-            indexDirectory = QnameCommand.getDefaultIndexPath(bamPath);
-        }
-        if (createDirectory(indexDirectory)) {
-            Path indexLevel1 = indexDirectory.resolve(IndexVersion.LATEST.fileName("index"));
-            Path indexLevel0 = indexDirectory.resolve(IndexVersion.LATEST.fileName("data"));
-            if (Files.exists(indexLevel1) || Files.exists(indexLevel0)) {
-                if (!force) {
-                    LOG.error("Index '{}' exists in '{}'.", IndexVersion.LATEST, indexDirectory);
-                    return false;
-                } else {
-                    LOG.error("Index '{}' exists in '{}', overwrite it", IndexVersion.LATEST, indexDirectory);
-                    return true;
-                }
-            }
-            return true;
-        }
-        return false;
-    }
-
-    private boolean createDirectory(Path path) {
-        if (!Files.exists(path)) {
-            try {
-                Files.createDirectory(path);
-            } catch (IOException e) {
-                LOG.error("Could not create directory: {}. Error : {}", path, e.getMessage());
-                return false;
-            }
-        }
-        return Files.isDirectory(path);
-    }
-}
-
-class QnameParam {
-    @Option(names = {"-n", "--name"}, description = "QNAME to search for")
-    String qname;
-    @Option(names = {"-f", "--file-name"}, description = "Path to a file containing QNAMEs to search for (separated by newline).")
-    Path qnamePath;
-
-    Set<String> getQnames() throws IllegalArgumentException {
-        if (qname != null) {
-            return Set.of(qname);
-        }
-        final Map<Boolean, Set<String>> qnames = readFile()
-                    .stream()
-                    .filter(StringUtils::isNotBlank)
-                    .collect(partitioningBy(this::isValidQname, toSet()));
-            final Set<String> invalidQnames = qnames.getOrDefault(false, Set.of());
-            if (!invalidQnames.isEmpty()) {
-                throw new IllegalArgumentException("File " + qnamePath + " contains invalid qnames : " + invalidQnames);
-            }
-            return qnames.get(true);
-    }
-
-    private List<String> readFile() {
-        try {
-            return Files.readAllLines(qnamePath);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("Could not read file: " + qnamePath, e);
-        }
-    }
-
-    private boolean isValidQname(String qName) {
-        try {
-            SAMSequenceRecord.validateSequenceName(qName);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-}
-
-@Command(name = "view",
-         header = "View records by QNAME",
-         description = "View records with the given QNAME from a BAM file. This command needs an index to be present" +
-                 " created with the `index` sub command.")
-class ViewCommand implements Callable<Integer> {
-    @Parameters(index = "0",
-            paramLabel = "bam-file",
-            description = "Path to the BAM file")
-    Path bamPath;
-
-    @ArgGroup(multiplicity = "1",
-              heading = "One of qname or file containing qnames")
-    QnameParam qnameParam;
-
-    @Option(names = {"-i", "--index-path"},
-            description = "Index directory")
-    Path indexDirectory;
-
-    @Option(names = {"-h", "--header"},
-            description = "Include header in SAM output",
-            defaultValue = "false")
-    boolean includeHeader;
-
-    @Option(names = {"-v", "--verbose"},
-            description = "Switch on verbose output",
-            defaultValue = "false")
-    boolean verbose;
-
-    @Override
-    public Integer call() throws Exception {
-        java.util.logging.Logger.getLogger("")
-                .setLevel(verbose ? java.util.logging.Level.ALL : java.util.logging.Level.SEVERE);
-
-        if (!Files.isRegularFile(bamPath)) {
-            System.err.println(bamPath + " not found.");
-            return -1;
-        }
-
-        if (indexDirectory == null) {
-            indexDirectory = QnameCommand.getDefaultIndexPath(bamPath);
-        }
-        if (!Files.isDirectory(indexDirectory)) {
-            System.err.println(indexDirectory + " not found.");
-            return -1;
-        }
-
-        Path indexLevel1 = indexDirectory.resolve(IndexVersion.LATEST.fileName("index"));
-        Path indexLevel0 = indexDirectory.resolve(IndexVersion.LATEST.fileName("data"));
-
-        if (!Files.exists(indexLevel1) || !Files.exists(indexLevel0)) {
-            System.err.printf("Index '%s' doesn't exists in '%s'.\n", IndexVersion.LATEST, indexDirectory);
-            return -1;
-        }
-
-        long start = System.nanoTime();
-
-        BamRecordReader bamRecordReader = new DefaultBamRecordReader(new SamtoolsBasedParser());
-        KeyPointerReader qnamePosReader = new KeyPointerReader();
-        SamPrintingHandler handler = new SamPrintingHandler(System.out, includeHeader);
-        QnameSearcher searcher = new QnameSearcher(qnamePosReader, bamRecordReader, handler);
-
-        final Set<String> qnames;
-        try {
-            qnames = qnameParam.getQnames();
-        } catch (IllegalArgumentException e) {
-            LOG.error(e.getMessage());
-            return -1;
-        }
-        searcher.search(bamPath, indexDirectory, qnames);
-        handler.finish();
-
-        long finish = System.nanoTime();
-
-        LOG.info("Search completed in " + (finish - start) / 1000_000 + "ms");
-        return 0;
+        return new String[] { "Version: " + BASE_VERSION };
     }
 }
